+++ sdk-java/src/main/java/ly/count/sdk/java/internal/SessionImpl.java
@@ -118,6 +118,9 @@ public class SessionImpl implements Session, Storable, EventImpl.EventRecorder {
         }
 
         this.consents = SDKCore.instance.consents;
+        if (config.isAutoSendUserPropertiesOnSessions() && config.sdk.userProfile() != null) {
+            config.sdk.module(ModuleUserProfile.class).saveInternal();
+        }
 
         if (pushOnChange) {
             Storage.pushAsync(config, this);
@@ -157,6 +160,9 @@ public class SessionImpl implements Session, Storable, EventImpl.EventRecorder {
         }
 
         this.consents = SDKCore.instance.consents;
+        if (config.isAutoSendUserPropertiesOnSessions() && config.sdk.userProfile() != null) {
+            config.sdk.module(ModuleUserProfile.class).saveInternal();
+        }
 
         Long duration = updateDuration(now);
 
@@ -192,6 +198,9 @@ public class SessionImpl implements Session, Storable, EventImpl.EventRecorder {
         ended = now == null ? System.nanoTime() : now;
 
         this.consents = SDKCore.instance.consents;
+        if (config.isAutoSendUserPropertiesOnSessions() && config.sdk.userProfile() != null) {
+            config.sdk.module(ModuleUserProfile.class).saveInternal();
+        }
 
         if (currentView != null) {
             currentView.stop(true);